chore: Update Textarea stories to use index approach
