+++ packages/react-components/react-textarea/src/stories/Textarea/TextareaDefault.stories.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import { Textarea } from '../Textarea';
 import { useId } from '@fluentui/react-utilities';
+import { Textarea } from '@fluentui/react-textarea';
+import type { TextareaProps } from '@fluentui/react-textarea';
 import { Label } from '@fluentui/react-label';
-import type { TextareaProps } from '../Textarea';
 import { makeStyles } from '@griffel/react';
 import { tokens } from '@fluentui/react-theme';
 
@@ -16,23 +16,14 @@ const useStyles = makeStyles({
   },
 });
 
-export const Controlled = () => {
+export const Default = (props: Partial<TextareaProps>) => {
   const textareaId = useId('textarea');
   const styles = useStyles();
-  const [value, setValue] = React.useState('initial value');
-
-  const onChange: TextareaProps['onChange'] = (ev, data) => {
-    if (data.value.length <= 50) {
-      setValue(data.value);
-    }
-  };
 
   return (
     <div className={styles.base}>
-      <Label htmlFor={textareaId} style={{ display: 'block' }}>
-        Controlled Textarea limiting the value to 50 characters.
-      </Label>
-      <Textarea value={value} onChange={onChange} id={textareaId} />
+      <Label htmlFor={textareaId}>Default Textarea</Label>
+      <Textarea id={textareaId} {...props} />
     </div>
   );
 };
