--- packages/react-components/react-textarea/src/stories/TextareaDisabled.stories.tsx
@@ -1,28 +1,0 @@
-import * as React from 'react';
-import { useId } from '@fluentui/react-utilities';
-import { Label } from '@fluentui/react-label';
-import { Textarea } from '../Textarea';
-import { makeStyles } from '@griffel/react';
-import { tokens } from '@fluentui/react-theme';
-
-const useStyles = makeStyles({
-  base: {
-    display: 'flex',
-    flexDirection: 'column',
-    '& > label': {
-      marginBottom: tokens.spacingVerticalMNudge,
-    },
-  },
-});
-
-export const Disabled = () => {
-  const disabledId = useId('textarea-disabled');
-  const styles = useStyles();
-
-  return (
-    <div className={styles.base}>
-      <Label htmlFor={disabledId}>Disabled Textarea.</Label>
-      <Textarea id={disabledId} disabled />
-    </div>
-  );
-};
